requirements.txt
Django==1.6.2
South==0.8.4
wsgiref==0.1.2
Pillow==2.4.0
easy-thumbnails==1.5
django-jfu==2.0.1
